Several fixes. Unchecked return pointers and some other conditions. Fixed problem with request handler processes on both master and slave that caused the recently spawned process not to die when it should. Some other changes: "Use the source, Luke..."

+++ libdrqueue/semaphores.c
@@ -33,8 +33,9 @@ void semaphore_lock (int64_t semid) {
   op.sem_flg = SEM_UNDO;
 
   /*  fprintf (stderr,"Locking... semval: %i semid: %i\n",semctl (semid,0,GETVAL),semid); */
-  if (semop(semid,&op,1) == -1) {
+  if (semop((int)semid,&op,1) == -1) {
     perror ("semaphore_lock");
+    // FIXME: clean exit instead of kill.
     kill(0,SIGINT);
   }
   /*  fprintf (stderr,"Locked !!! semval: %i semid: %i\n",semctl (semid,0,GETVAL),semid); */
@@ -47,8 +48,9 @@ void semaphore_release (int64_t semid) {
   op.sem_num = 0;
   op.sem_op = 1;
   op.sem_flg = SEM_UNDO; /*  SEM_UNDO; */
-  if (semop(semid,&op,1) == -1) {
+  if (semop((int)semid,&op,1) == -1) {
     perror ("semaphore_release");
+    // FIXME: clean exit instead of kill.
     kill(0,SIGINT);
   }
   /*  fprintf (stderr,"Unlocked !!! semval: %i semid: %i\n",semctl (semid,0,GETVAL),semid); */
